pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "auto-video-agent"
version = "0.3.0"
description = "AI-powered CLI tool that generates short-form marketing videos for auto repair shops"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "click>=8.1.0",
    "pyyaml>=6.0",
    "edge-tts>=6.1.0",
    "ffmpeg-python>=0.2.0",
    "Pillow>=10.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[project.scripts]
auto-video-agent = "auto_video_agent.cli:main"

[tool.setuptools.packages.find]
include = ["auto_video_agent*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
